Use a PTY instead of a pipe when possible
ugoren pointed out that the competition rules specifically stated using
fflush() in submissions is unnecessary if a newline follows each letter,
yet the standard I/O library appears to not consider a pipe to be an
"interactive" device. This commit fixes the tester's problem by replacing
the pipe with a master/slave pair of PTY devices when the shell command
to invoke the program under test is provided as a command line argument.